# backend/openapi.yml
openapi: 3.1.0
info:
  title: SMSGang API
  version: 1.0.0
  description: |
    The SMSGang REST API enables purchasing temporary virtual phone numbers for
    SMS verification on 300+ global platforms. All monetary values are in **Nigerian Naira (₦)**.

    ## Authentication
    The API uses **Laravel Sanctum** bearer tokens. After registering or logging in,
    include the token in every protected request via the `Authorization` header:
    ```
    Authorization: Bearer {token}
    ```

    ## Rate Limits
    | Scope | Limit |
    |---|---|
    | `auth` | 10 req / min |
    | `api` | 60 req / min |
    | `buy` | 5 purchases / min |
    | `sms-check` | 10 checks / min |
    | `webhook` | 30 req / min |

  contact:
    name: SMSGang Support
  license:
    name: Proprietary
servers:
  - url: https://api.smsgang.ng/api
    description: Production
  - url: http://localhost:8000/api
    description: Local Development
...
paths:
...
  /services/{service}/countries:
    get:
      tags: [Catalog]
      summary: Get available countries and live prices for a service
      operationId: catalog.service.countries
      parameters:
        - $ref: '#/components/parameters/ServiceSlug'
      responses:
        '200':
          description: Countries with real-time prices
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceCountryPrice'
        '404':
          $ref: '#/components/responses/NotFound'
  /countries/{country}/services:
    get:
      tags: [Catalog]
      summary: Get available services for a country
      operationId: catalog.country.services
      parameters:
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Services available in the given country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCollection'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ACTIVATIONS ──────────────────────────────
...
  /activations/{order}/verify-payment:
    post:
      tags: [Activations]
      summary: Verify payment and assign a phone number
      description: |
        Call this endpoint after the user returns from the payment gateway.
        If payment is confirmed the phone number is instantly assigned.
      operationId: activations.verifyPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Payment confirmed and number assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment verified. Number assigned.
                  activation:
                    $ref: '#/components/schemas/Activation'
        '402':
          description: Payment not yet confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment not confirmed yet.
                  payment_status:
                    type: string
                    example: pending
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /activations/{activation}/check-sms:
    get:
      tags: [Activations]
      summary: Poll for the SMS verification code
      description: Fetches the latest SMS from the provider and updates the activation status.
      operationId: activations.checkSms
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivationId'
      responses:
        '200':
          description: Latest activation state with SMS code if received
          content:
            application/json:
              schema:
                type: object
                properties:
                  activation:
                    $ref: '#/components/schemas/Activation'
                  sms_code:
                    type: string
                    nullable: true
                    example: '482910'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/services:
    get:
      tags: [Admin – Services]
      summary: List all services
      operationId: admin.services.index
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Full service list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCollection'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /webhooks/lendoverify:
    post:
      tags: [Webhooks]
      summary: Lendoverify payment webhook
      description: |
        Receives payment status events from the Lendoverify gateway.
        This endpoint is **public** (no auth) but validates the event payload.
        Do not call this endpoint manually.
      operationId: webhooks.lendoverify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid payload

# ─────────────────────────────────────────────
#  COMPONENTS
# ─────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum personal access token

  # ── PARAMETERS ──────────────────────────────
  parameters:
...
    ServiceSlug:
      name: service
      in: path
      required: true
      schema:
        type: string
      description: Service slug (e.g. `whatsapp`)
...
    CountryCode:
      name: country
      in: path
      required: true
      schema:
        type: string
      description: Country ISO code or slug (e.g. `NG`)
    ActivationId:
      name: activation
      in: path
      required: true
      schema:
        type: integer
      description: Activation ID
    OrderId:
      name: order
      in: path
      required: true
      schema:
        type: integer
      description: Order ID
...
  schemas:
...
    ActivationStatus:
      type: string
      enum:
        - requested
        - number_received
        - waiting_sms
        - sms_received
        - completed
        - expired
        - cancelled
...
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: WhatsApp
        slug:
          type: string
          example: whatsapp
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    Country:
      type: object
      properties:
        id:
          type: integer
          example: 5
        name:
          type: string
          example: Nigeria
        code:
          type: string
          example: NG
        flag:
          type: string
          example: 🇳🇬
        is_active:
          type: boolean
          example: true
...
    ServiceCountryPrice:
      type: object
      description: Live price entry as returned by /services/{service}/countries
      properties:
        country:
          $ref: '#/components/schemas/Country'
        cost_ngn:
          type: number
          format: float
          description: Final price in NGN including exchange rate and global markup
          example: 350.00
        available_count:
          type: integer
          example: 4821
...
    Activation:
      type: object
      properties:
        id:
          type: integer
          example: 55
        order_id:
          type: integer
          example: 100
        phone_number:
          type: string
          example: '+79161234567'
        sms_code:
          type: string
          nullable: true
          example: '482910'
        status:
          $ref: '#/components/schemas/ActivationStatus'
        provider:
          type: string
          example: 5sim
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        service:
          $ref: '#/components/schemas/Service'
        country:
          $ref: '#/components/schemas/Country'
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email has already been taken.

    # ---- Paginated Collections ----------------
...
    ServiceCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
...
  responses:
...
    Forbidden:
      description: Forbidden — insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: This action is unauthorized.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Not found.
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'